requirements.txt
openai==1.86.0
langchain==0.2.17
langchain-openai==0.1.25
python-dotenv==1.0.1